+++ test/parallel/parallel.status
@@ -40,8 +40,6 @@ test-async-hooks-http-parser-destroy: PASS,FLAKY
 # https://github.com/nodejs/node/pull/31178
 test-crypto-dh-stateless: SKIP
 test-crypto-keygen: SKIP
-# https://github.com/nodejs/node/issues/33796
-test-fs-stream-construct: PASS,FLAKY
 
 [$system==solaris] # Also applies to SmartOS
